requirements.txt
requests==2.31.0
PyQt5==5.15.10
python-dotenv==1.0.0
